Split the logic of showing the participantsbutton
